cue/load/internal/fileprocessor: factor out of cue/load
This means we will be able to move some modules-related logic
into a separate package to cue/load. It also makes the internal API
boundaries a bit clearer.
