+++ cue/load/match.go
@@ -15,14 +15,10 @@
 package load
 
 import (
-	"io/ioutil"
-	"path/filepath"
-	"regexp"
 	"strings"
 
 	"cuelang.org/go/cue/build"
 	"cuelang.org/go/cue/errors"
-	"cuelang.org/go/cue/token"
 )
 
 // A match represents the result of matching a single package pattern.
@@ -41,134 +37,6 @@ type excludeError struct {
 }
 
 func (e excludeError) Is(err error) bool { return err == errExclude }
-
-// matchFile determines whether the file with the given name in the given directory
-// should be included in the package being constructed.
-// It returns the data read from the file.
-// If returnImports is true and name denotes a CUE file, matchFile reads
-// until the end of the imports (and returns that data) even though it only
-// considers text until the first non-comment.
-// If allTags is non-nil, matchFile records any encountered build tag
-// by setting allTags[tag] = true.
-func matchFile(cfg *fileProcessorConfig, file *build.File, returnImports, allFiles bool, allTags map[string]bool) (match bool, data []byte, err errors.Error) {
-	if fi := cfg.fileSystem.getOverlay(file.Filename); fi != nil {
-		if fi.file != nil {
-			file.Source = fi.file
-		} else {
-			file.Source = fi.contents
-		}
-	}
-
-	if file.Encoding != build.CUE {
-		return false, nil, nil // not a CUE file, don't record.
-	}
-
-	if file.Filename == "-" {
-		b, err2 := ioutil.ReadAll(cfg.stdin())
-		if err2 != nil {
-			err = errors.Newf(token.NoPos, "read stdin: %v", err)
-			return
-		}
-		file.Source = b
-		return true, b, nil // don't check shouldBuild for stdin
-	}
-
-	name := filepath.Base(file.Filename)
-	if !cfg.filesMode && strings.HasPrefix(name, ".") {
-		return false, nil, &excludeError{
-			errors.Newf(token.NoPos, "filename starts with a '.'"),
-		}
-	}
-
-	if strings.HasPrefix(name, "_") {
-		return false, nil, &excludeError{
-			errors.Newf(token.NoPos, "filename starts with a '_"),
-		}
-	}
-
-	f, err := cfg.fileSystem.openFile(file.Filename)
-	if err != nil {
-		return false, nil, err
-	}
-
-	data, err = readImports(f, false, nil)
-	f.Close()
-	if err != nil {
-		return false, nil,
-			errors.Newf(token.NoPos, "read %s: %v", file.Filename, err)
-	}
-
-	return true, data, nil
-}
-
-// treeCanMatchPattern(pattern)(name) reports whether
-// name or children of name can possibly match pattern.
-// Pattern is the same limited glob accepted by matchPattern.
-func treeCanMatchPattern(pattern string) func(name string) bool {
-	wildCard := false
-	if i := strings.Index(pattern, "..."); i >= 0 {
-		wildCard = true
-		pattern = pattern[:i]
-	}
-	return func(name string) bool {
-		return len(name) <= len(pattern) && hasPathPrefix(pattern, name) ||
-			wildCard && strings.HasPrefix(name, pattern)
-	}
-}
-
-// matchPattern(pattern)(name) reports whether
-// name matches pattern. Pattern is a limited glob
-// pattern in which '...' means 'any string' and there
-// is no other special syntax.
-// Unfortunately, there are two special cases. Quoting "go help packages":
-//
-// First, /... at the end of the pattern can match an empty string,
-// so that net/... matches both net and packages in its subdirectories, like net/http.
-// Second, any slash-separted pattern element containing a wildcard never
-// participates in a match of the "vendor" element in the path of a vendored
-// package, so that ./... does not match packages in subdirectories of
-// ./vendor or ./mycode/vendor, but ./vendor/... and ./mycode/vendor/... do.
-// Note, however, that a directory named vendor that itself contains code
-// is not a vendored package: cmd/vendor would be a command named vendor,
-// and the pattern cmd/... matches it.
-func matchPattern(pattern string) func(name string) bool {
-	// Convert pattern to regular expression.
-	// The strategy for the trailing /... is to nest it in an explicit ? expression.
-	// The strategy for the vendor exclusion is to change the unmatchable
-	// vendor strings to a disallowed code point (vendorChar) and to use
-	// "(anything but that codepoint)*" as the implementation of the ... wildcard.
-	// This is a bit complicated but the obvious alternative,
-	// namely a hand-written search like in most shell glob matchers,
-	// is too easy to make accidentally exponential.
-	// Using package regexp guarantees linear-time matching.
-
-	const vendorChar = "\x00"
-
-	if strings.Contains(pattern, vendorChar) {
-		return func(name string) bool { return false }
-	}
-
-	re := regexp.QuoteMeta(pattern)
-	re = replaceVendor(re, vendorChar)
-	switch {
-	case strings.HasSuffix(re, `/`+vendorChar+`/\.\.\.`):
-		re = strings.TrimSuffix(re, `/`+vendorChar+`/\.\.\.`) + `(/vendor|/` + vendorChar + `/\.\.\.)`
-	case re == vendorChar+`/\.\.\.`:
-		re = `(/vendor|/` + vendorChar + `/\.\.\.)`
-	case strings.HasSuffix(re, `/\.\.\.`):
-		re = strings.TrimSuffix(re, `/\.\.\.`) + `(/\.\.\.)?`
-	}
-	re = strings.Replace(re, `\.\.\.`, `[^`+vendorChar+`]*`, -1)
-
-	reg := regexp.MustCompile(`^` + re + `$`)
-
-	return func(name string) bool {
-		if strings.Contains(name, vendorChar) {
-			return false
-		}
-		return reg.MatchString(replaceVendor(name, vendorChar))
-	}
-}
 
 // replaceVendor returns the result of replacing
 // non-trailing vendor path elements in x with repl.
